Began working on a new 0.16 update and a new NBT library.
Why do we need to do this?

I decided to start over with a new branch as a lot has changed in Voxelwind since my last 0.16-related work in late August.

As for the NBT library, this is because Mojang introduced a new NBT encoding in MCPE 0.16. This encoding uses VarInts instead of regular types. None of the other NBT libraries are active, so I'm forced to create a new library. This means I can do some things I've wanted to do for a while (immutable tag instances!)